backend/requirements.txt
flask
paho-mqtt
flask-cors
flask-pymongo
pandas
prophet
